Added modify copropiedad component.

+++ PropiedadHorizonta.Core/DTO/CopropiedadesDto.cs
@@ -3,13 +3,13 @@ namespace PropiedadHorizontal.Core.DTO
 {
     public class CopropiedadesDto
     {
-        public int IdCopropiedad { get; set; }
+        public long IdCopropiedad { get; set; }
         public string NombreCopropiedad { get; set; }
         public string NitPropiedadHorizontal { get; set; }
         public decimal? Indice { get; set; }
         public decimal? Area { get; set; }
         public string IdDocumentoCopropietario { get; set; }
-        public decimal? IdTipoCopropiedad { get; set; }
+        public int? IdTipoCopropiedad { get; set; }
         public string DescripcionTipoCopropiedad { get; set; }
         public string NombrePropiedadHorizontal { get; set; }
         public string NombrePropietario { get; set; }
